move device related params from disk to device
