+++ hazelcast-spring-tests/src/test/resources/com/hazelcast/spring/fullConfig-applicationContext-hazelcast.xml
@@ -473,10 +473,7 @@
                     <hz:memory-tier>
                         <hz:capacity>128 MB</hz:capacity>
                     </hz:memory-tier>
-                    <hz:disk-tier enabled="true">
-                        <hz:base-dir>tiered-store_base_dir</hz:base-dir>
-                        <hz:block-size>512</hz:block-size>
-                    </hz:disk-tier>
+                    <hz:disk-tier enabled="true"/>
                 </hz:tiered-store>
             </hz:map>
 
@@ -1064,6 +1061,7 @@
             <hz:device>
                 <hz:device-name>tiered_store_device</hz:device-name>
                 <hz:base-dir>/dev/devices/tiered_store_device</hz:base-dir>
+                <hz:block-size>8192</hz:block-size>
             </hz:device>
 
             <hz:auditlog enabled="false" factory-class-name="com.acme.AuditlogToSyslogFactory">
