+++ hazelcast/src/main/java/com/hazelcast/config/DiskTierConfig.java
@@ -21,9 +21,7 @@ import com.hazelcast.nio.ObjectDataInput;
 import com.hazelcast.nio.ObjectDataOutput;
 import com.hazelcast.nio.serialization.IdentifiedDataSerializable;
 
-import java.io.File;
 import java.io.IOException;
-import java.util.Objects;
 
 /**
  * Disk tier configuration of Tiered-Store.
@@ -32,19 +30,7 @@ import java.util.Objects;
  */
 public class DiskTierConfig implements IdentifiedDataSerializable {
 
-    /**
-     * Default base directory for the tiered-store.
-     */
-    public static final String DEFAULT_TIERED_STORE_BASE_DIR = "tiered-store";
-
-    /**
-     * Default block/sector size in bytes.
-     */
-    public static final int DEFAULT_BLOCK_SIZE_IN_BYTES = 4096;
-
     private boolean enabled;
-    private File baseDir = new File(DEFAULT_TIERED_STORE_BASE_DIR).getAbsoluteFile();
-    private int blockSize = DEFAULT_BLOCK_SIZE_IN_BYTES;
 
     public DiskTierConfig() {
 
@@ -52,8 +38,6 @@ public class DiskTierConfig implements IdentifiedDataSerializable {
 
     public DiskTierConfig(DiskTierConfig diskTierConfig) {
         enabled = diskTierConfig.isEnabled();
-        baseDir = diskTierConfig.getBaseDir();
-        blockSize = diskTierConfig.getBlockSize();
     }
 
     /**
@@ -76,57 +60,14 @@ public class DiskTierConfig implements IdentifiedDataSerializable {
         return this;
     }
 
-    /**
-     * Base directory for this disk tier.
-     * Can be an absolute or relative path to the node startup directory.
-     */
-    public File getBaseDir() {
-        return baseDir;
-    }
-
-    /**
-     * Sets the base directory for this disk tier.
-     *
-     * @param baseDir base directory.
-     * @return this DiskTierConfig
-     */
-    public DiskTierConfig setBaseDir(File baseDir) {
-        this.baseDir = baseDir;
-        return this;
-    }
-
-    /**
-     * Returns disk block/sector size in bytes.
-     *
-     * @return block size
-     */
-    public int getBlockSize() {
-        return blockSize;
-    }
-
-    /**
-     * Sets the disk block/sector size in bytes.
-     * @param blockSize block size.
-     * @return this DiskTierConfig
-     */
-    public DiskTierConfig setBlockSize(int blockSize) {
-        this.blockSize = blockSize;
-        return this;
-    }
-
-
     @Override
     public void writeData(ObjectDataOutput out) throws IOException {
         out.writeBoolean(enabled);
-        out.writeString(baseDir.getAbsolutePath());
-        out.writeInt(blockSize);
     }
 
     @Override
     public void readData(ObjectDataInput in) throws IOException {
         enabled = in.readBoolean();
-        baseDir = new File(in.readString()).getAbsoluteFile();
-        blockSize = in.readInt();
     }
 
     @Override
@@ -150,29 +91,18 @@ public class DiskTierConfig implements IdentifiedDataSerializable {
 
         DiskTierConfig that = (DiskTierConfig) o;
 
-        if (enabled != that.enabled) {
-            return false;
-        }
-        if (blockSize != that.blockSize) {
-            return false;
-        }
-        return Objects.equals(baseDir, that.baseDir);
+        return enabled == that.enabled;
     }
 
     @Override
     public final int hashCode() {
-        int result = (enabled ? 1 : 0);
-        result = 31 * result + (baseDir != null ? baseDir.hashCode() : 0);
-        result = 31 * result + blockSize;
-        return result;
+        return (enabled ? 1 : 0);
     }
 
     @Override
     public String toString() {
         return "DiskTierConfig{"
                 + "enabled=" + enabled
-                + ", baseDir=" + baseDir
-                + ", blockSize=" + blockSize
                 + '}';
     }
 }
