+++ hazelcast/src/main/resources/hazelcast-config-5.1.xsd
@@ -4099,19 +4099,31 @@
             </xs:documentation>
         </xs:annotation>
         <xs:all>
-            <xs:element name="device-name" type="xs:string">
+            <xs:element name="device-name" type="xs:string" minOccurs="0" default="local-disk">
                 <xs:annotation>
                     <xs:documentation>
                         Device name.
                     </xs:documentation>
                 </xs:annotation>
             </xs:element>
-            <xs:element name="base-dir" type="xs:string" minOccurs="0" default="tiered-store-device">
+            <xs:element name="base-dir" type="xs:string" minOccurs="0" default="tiered-store">
                 <xs:annotation>
                     <xs:documentation>
-                        Base directory in the device that will be used for tiered-store.
+                        Device base directory that will be used for tiered-store.
                     </xs:documentation>
                 </xs:annotation>
+            </xs:element>
+            <xs:element name="block-size" minOccurs="0" default="4096">
+                <xs:annotation>
+                    <xs:documentation>
+                        Device block/sector size in bytes. The default is 4096. The minimum is 512.
+                    </xs:documentation>
+                </xs:annotation>
+                <xs:simpleType>
+                    <xs:restriction base="xs:unsignedInt">
+                        <xs:minInclusive value="512"/>
+                    </xs:restriction>
+                </xs:simpleType>
             </xs:element>
         </xs:all>
     </xs:complexType>
@@ -4146,27 +4158,6 @@
                 Defines configuration for the disk tier of a tiered store.
             </xs:documentation>
         </xs:annotation>
-        <xs:all>
-            <xs:element name="base-dir" type="xs:string" minOccurs="0" default="tiered-store">
-                <xs:annotation>
-                    <xs:documentation>
-                        Base directory for storing files.
-                    </xs:documentation>
-                </xs:annotation>
-            </xs:element>
-            <xs:element name="block-size" minOccurs="0" default="4096">
-                <xs:annotation>
-                    <xs:documentation>
-                        Disk block/sector size in bytes. The default is 4096. The minimum is 512.
-                    </xs:documentation>
-                </xs:annotation>
-                <xs:simpleType>
-                    <xs:restriction base="xs:unsignedInt">
-                        <xs:minInclusive value="512"/>
-                    </xs:restriction>
-                </xs:simpleType>
-            </xs:element>
-        </xs:all>
         <xs:attribute name="enabled" type="xs:boolean" use="required">
             <xs:annotation>
                 <xs:documentation>
