+++ hazelcast/src/test/resources/com/hazelcast/config/yaml/testcases/member/tiered-store/multiple-failures.json
@@ -2,7 +2,8 @@
   "instance": {
     "hazelcast": {
       "device": {
-        "base-dir": "tiered_store_base_dir"
+        "base-dir": 1111,
+        "block-size": 511
       },
       "map": {
         "my-map-0": {
@@ -19,11 +20,6 @@
             "enabled": true,
             "memory-tier": {
               "capacity": -1
-            },
-            "disk-tier": {
-              "enabled": true,
-              "base-dir": 1111,
-              "block-size": 511
             }
           }
         },
@@ -45,9 +41,23 @@
       {
         "schemaLocation": "#/definitions/Device",
         "pointerToViolation": "#/hazelcast/device",
-        "causingExceptions": [],
-        "keyword": "required",
-        "message": "required key [device-name] not found"
+        "causingExceptions": [
+          {
+            "schemaLocation": "#/definitions/Device/properties/base-dir",
+            "pointerToViolation": "#/hazelcast/device/base-dir",
+            "causingExceptions": [],
+            "keyword": "type",
+            "message": "expected type: String, found: Integer"
+          },
+          {
+            "schemaLocation": "#/definitions/Device/properties/block-size",
+            "pointerToViolation": "#/hazelcast/device/block-size",
+            "causingExceptions": [],
+            "keyword": "minimum",
+            "message": "511 is not greater or equal to 512"
+          }
+        ],
+        "message": "2 schema violations found"
       },
       {
         "schemaLocation": "#/definitions/Map",
@@ -61,46 +71,18 @@
             "message": "required key [enabled] not found"
           },
           {
-            "schemaLocation": "#/definitions/Map/additionalProperties/properties/tiered-store",
-            "pointerToViolation": "#/hazelcast/map/my-map-2/tiered-store",
-            "causingExceptions": [
-              {
-                "schemaLocation": "#/definitions/Map/additionalProperties/properties/tiered-store/properties/disk-tier",
-                "pointerToViolation": "#/hazelcast/map/my-map-2/tiered-store/disk-tier",
-                "causingExceptions": [
-                  {
-                    "schemaLocation": "#/definitions/Map/additionalProperties/properties/tiered-store/properties/disk-tier/properties/base-dir",
-                    "pointerToViolation": "#/hazelcast/map/my-map-2/tiered-store/disk-tier/base-dir",
-                    "causingExceptions": [],
-                    "keyword": "type",
-                    "message": "expected type: String, found: Integer"
-                  },
-                  {
-                    "schemaLocation": "#/definitions/Map/additionalProperties/properties/tiered-store/properties/disk-tier/properties/block-size",
-                    "pointerToViolation": "#/hazelcast/map/my-map-2/tiered-store/disk-tier/block-size",
-                    "causingExceptions": [],
-                    "keyword": "minimum",
-                    "message": "511 is not greater or equal to 512"
-                  }
-                ],
-                "message": "2 schema violations found"
-              },
-              {
-                "schemaLocation": "#/definitions/Map/additionalProperties/properties/tiered-store/properties/memory-tier/properties/capacity",
-                "pointerToViolation": "#/hazelcast/map/my-map-2/tiered-store/memory-tier/capacity",
-                "causingExceptions": [],
-                "keyword": "type",
-                "message": "expected type: String, found: Integer"
-              }
-            ],
-            "message": "3 schema violations found"
-          },
-          {
             "schemaLocation": "#/definitions/Map/additionalProperties/properties/tiered-store/properties/disk-tier",
             "pointerToViolation": "#/hazelcast/map/my-map-1/tiered-store/disk-tier",
             "causingExceptions": [],
             "keyword": "required",
             "message": "required key [enabled] not found"
+          },
+          {
+            "schemaLocation": "#/definitions/Map/additionalProperties/properties/tiered-store/properties/memory-tier/properties/capacity",
+            "pointerToViolation": "#/hazelcast/map/my-map-2/tiered-store/memory-tier/capacity",
+            "causingExceptions": [],
+            "keyword": "type",
+            "message": "expected type: String, found: Integer"
           },
           {
             "schemaLocation": "#/definitions/Map/additionalProperties/properties/tiered-store/properties/memory-tier/properties/capacity",
@@ -110,9 +92,9 @@
             "message": "string [01 KB] does not match pattern ^[1-9]([0-9]*) (B|KB|MB|GB)$"
           }
         ],
-        "message": "6 schema violations found"
+        "message": "4 schema violations found"
       }
     ],
-    "message": "7 schema violations found"
+    "message": "6 schema violations found"
   }
 }
